admin/requirements.txt
Django==3.1.6
djangorestframework==3.12.2
psycopg2==2.8.6
pika==1.1.0
django-cors-headers